requirements.txt
Django~=4.0.4
Pillow
requests~=2.26.0
djangorestframework~=3.13.1